usbuart: Moved usbuart_change_dma_tx_buf() into aux_serial.c
